+++ src/machinable/registration/registration.py
@@ -89,6 +89,21 @@ class Registration:
         """
         pass
 
+    def on_resolve_vendor(self, name, config, target):
+        """Event triggered when vendor is resolved
+
+        machinable attempts to fetch any vendor directories that
+        are registered under the '+'-section in the machinable.yaml
+
+        # Arguments
+        name: The name of the vendor
+        config: The vendor configuration from the machinable.yaml
+        target: The target directory (may or may not exists yet)
+
+        Return False to prevent the default automatic resolution
+        """
+        pass
+
     def default_resources(self, engine, component, components):
         """Allows to specify global default resources"""
         return None